Bottom up branch clean up

+++ core/bottom_up/src/main/java/com/gitee/drinkjava2/frog/brain/Cuboid.java
@@ -10,52 +10,50 @@
  */
 package com.gitee.drinkjava2.frog.brain;
 
-import com.gitee.drinkjava2.frog.Frog;
-
 /**
  * Cuboid represents a rectangular prism 3d zone in brain
  * 
- * Cuboid是一个长方体，通常用来表示脑内器官的形状
+ * Cuboid是一个长方体，通常用来表示脑内器官的形状。另一个功能类似的形装是Cone锥形体。
  * 
  * @author Yong Zhu
  * @since 2.0.2
  */
 @SuppressWarnings("all")
 public class Cuboid implements Shape {
-    private static final long serialVersionUID = 1L;
+	private static final long serialVersionUID = 1L;
 
-    public int x;// x,y,z是长方体的左下角坐标
-    public int y;
-    public int z;
-    public int xe;// xe,ye,ze分别是长方体三边长
-    public int ye;
-    public int ze;
+	public float x;// x,y,z是长方体的左下角坐标
+	public float y;
+	public float z;
+	public float xe;// xe,ye,ze分别是长方体三边长
+	public float ye;
+	public float ze;
 
-    public Cuboid() {
-        // 空构造器不能省
-    }
+	public Cuboid() {
+		// 空构造器不能省
+	}
 
-    public Cuboid(int x, int y, int z, int xe, int ye, int ze) {// 用x,y,z,r来构造
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.xe = xe;
-        this.ye = ye;
-        this.ze = ze;
-    }
+	public Cuboid(float x, float y, float z, float xe, float ye, float ze) {// 用x,y,z,r来构造
+		this.x = x;
+		this.y = y;
+		this.z = z;
+		this.xe = xe;
+		this.ye = ye;
+		this.ze = ze;
+	}
 
-    public Cuboid(Cuboid c) {// 用另一个Cuboid来构造
-        this.x = c.x;
-        this.y = c.y;
-        this.z = c.z;
-        this.xe = c.xe;
-        this.ye = c.ye;
-        this.ze = c.ze;
-    }
+	public Cuboid(Cuboid c) {// 用另一个Cuboid来构造
+		this.x = c.x;
+		this.y = c.y;
+		this.z = c.z;
+		this.xe = c.xe;
+		this.ye = c.ye;
+		this.ze = c.ze;
+	}
 
-    @Override
-    public void drawOnBrainPicture(BrainPicture pic) {
-        pic.drawCuboid(this);
-    }
- 
+	@Override
+	public void drawOnBrainPicture(BrainPicture pic) {
+		pic.drawCuboid(this);
+	}
+
 }
